refactor(auth): create exception package
