+++ src/main/java/woowacourse/auth/exception/ForbiddenException.java
@@ -1,4 +1,4 @@
-package woowacourse.auth.application;
+package woowacourse.auth.exception;
 
 public class ForbiddenException extends RuntimeException {
